Revert .gitignore file
